+++ variables.tf
@@ -137,7 +137,12 @@ variable "linux_profile" {
 variable "service_cidr" {
   description = "CIDR used by kubernetes services (kubectl get svc)."
   type        = string
+}
 
+variable "outbound_type" {
+  description = "The outbound (egress) routing method which should be used for this Kubernetes Cluster. Possible values are `loadBalancer` and `userDefinedRouting`."
+  type        = string
+  default     = "loadBalancer"
 }
 
 variable "docker_bridge_cidr" {